+++ uamqp/client.py
@@ -194,7 +194,9 @@ class SendClient(AMQPClient):
             self._message_sender.open()
             return False
         elif self._message_sender._state == constants.MessageSenderState.Error:
-            raise errors.AMQPConnectionError("Message sender in error state.")
+            raise errors.AMQPConnectionError(
+                "Message Sender Client was unable to open. "
+                "Please confirm credentials and access permissions.")
         elif self._message_sender._state != constants.MessageSenderState.Open:
             return False
         return True
@@ -302,7 +304,9 @@ class ReceiveClient(AMQPClient):
             self._message_receiver.open(self)
             return False
         elif self._message_receiver._state == constants.MessageReceiverState.Error:
-            raise errors.AMQPConnectionError("Message receiver in error state.")
+            raise errors.AMQPConnectionError(
+                "Message Receiver Client was unable to open. "
+                "Please confirm credentials and access permissions.")
         elif self._message_receiver._state != constants.MessageReceiverState.Open:
             self._last_activity_timestamp = self._counter.get_current_ms()
             return False