+++ uamqp/__init__.py
@@ -43,18 +43,20 @@ def send_message(target, data, auth=None, debug=False):
 
 
 def receive_message(source, auth=None, timeout=0, debug=False):
-    received = receive_messages(source, auth=auth, batch_size=1, timeout=timeout, debug=debug)
+    received = receive_messages(source, auth=auth, max_batch_size=1, timeout=timeout, debug=debug)
     if received:
         return received[0]
     else:
         return None
 
 
-def receive_messages(source, auth=None, batch_size=None, timeout=0, debug=False, **kwargs):
-    if batch_size:
-        kwargs['prefetch'] = batch_size
-    with ReceiveClient(source, auth=auth, timeout=timeout, debug=debug, **kwargs) as receive_client:
-        return receive_client.receive_message_batch(batch_size=batch_size or receive_client._prefetch)
+def receive_messages(source, auth=None, max_batch_size=None, timeout=0, debug=False, **kwargs):
+    if max_batch_size:
+        kwargs['prefetch'] = max_batch_size
+    with ReceiveClient(source, auth=auth, debug=debug, **kwargs) as receive_client:
+        print("receiving batch", max_batch_size)
+        return receive_client.receive_message_batch(
+            max_batch_size=max_batch_size or receive_client._prefetch, timeout=timeout)
 
 
 class _Platform: