+++ build_many_linux_32bit.sh
@@ -4,7 +4,7 @@ set -e
 # To execute this script:
 # docker run --rm -v $PWD:/data pyca/cryptography-manylinux1:i686 /data/build_many_linux_32bit.sh
 
-export UAMQP_VERSION="0.1.0b1"
+export UAMQP_VERSION="0.1.0b2"
 
 export CPATH="/opt/pyca/cryptography/openssl/include"
 export LIBRARY_PATH="/opt/pyca/cryptography/openssl/lib"