+++ HISTORY.rst
@@ -8,10 +8,11 @@ Release History
 
 - Added message send retry.
 - Added timeouts and better error handling for management requests.
-- Improved connection and auth error handling.
+- Improved connection and auth error handling and error messages.
 - Fixed message annotations type.
 - SendClient.send_all_messages() now returns a list of message send statuses.
 - Fixed OpenSSL platform being initialized multiple times.
+- Fixed auto-refresh of SAS tokens.
 
 
 0.1.0b1 (2018-03-24)